release(1.6.4): wiki migrations — MySQL FK type compatibility
The two wiki migrations (20260617_000_knowledge,
20260617_001_knowledge_tag_map) declared FK reference columns as
`integer()` which maps to INT on MySQL. The referenced `id()` columns
map to BIGINT UNSIGNED. MySQL 8 refuses cross-type FKs and aborted the
schema apply with:

    SQLSTATE[HY000]: General error: 3780 Referencing column
    'category_id' and referenced column 'id' in foreign key constraint
    'fk_knowledge_pages_category_id' are incompatible.

This made every MySQL install since v1.5.6 unable to apply the wiki
tables. SQLite installs were unaffected — SQLite has only one INTEGER
affinity, so the same migration succeeded there.

Fix is type-narrow: swap `integer()` to `bigInteger()` for the four FK
columns (category_id, author_id, page_id, snapshot_by in 000;
page_id, tag_id in 001). On SQLite this is a no-op (same affinity).
On MySQL the column type now matches the referenced id() column.

Editing applied migration bodies normally violates the "append-only"
rule in CLAUDE.md, but the rule is about not breaking the apply log:
SQLite installs already ran fine and stay byte-equivalent, while
MySQL installs never completed the apply (so schema_migrations doesn't
have these rows yet — re-running with the corrected file just succeeds
the first time).

Tests 344/344 green (SQLite). MySQL CI job should now apply.

Note: the e2e jobs (`api-tokens admin tokens view`, `polls voter
table`) are still red, but those have been failing on `main` since
commit 0c1e934 (Jun 4, before any wiki work) — unrelated regressions
in the API-tokens admin view and the polls voter flow. Will tackle
separately.

+++ system/Database/migrations/20260617_000_knowledge.php
@@ -34,8 +34,11 @@ return function (Schema $schema): void {
         // show via Parsedown + HtmlSanitizer (cached at request scope by
         // the renderer if needed; for v1 we re-parse each load).
         $t->text('body_md')->nullable();
-        $t->integer('category_id')->nullable();
-        $t->integer('author_id');
+        // bigInteger (not integer) so the FK references line up on MySQL —
+        // id() maps to BIGINT UNSIGNED, and MySQL 8 refuses cross-type FKs.
+        // On SQLite both map to INTEGER affinity so there's no data drift.
+        $t->bigInteger('category_id')->nullable();
+        $t->bigInteger('author_id');
         $t->timestamp('created_at');
         $t->timestamp('updated_at');
         $t->foreign('category_id')->references('id')->on('knowledge_categories')->onDelete('SET NULL');
@@ -47,7 +50,7 @@ return function (Schema $schema): void {
 
     $schema->createTable('knowledge_page_versions', function (Blueprint $t) {
         $t->id();
-        $t->integer('page_id');
+        $t->bigInteger('page_id');
         // Snapshot of title + body_md at the moment "Save snapshot" was
         // pressed. Append-only — versions are a read journal, never
         // restorable to head (per design discussion 2026-06-17).
@@ -57,7 +60,7 @@ return function (Schema $schema): void {
         // "Approved by Legal"). NULL is fine; renders as a generic
         // "Snapshot #N" in the journal.
         $t->string('note', 280)->nullable();
-        $t->integer('snapshot_by');
+        $t->bigInteger('snapshot_by');
         $t->timestamp('snapshot_at');
         $t->foreign('page_id')->references('id')->on('knowledge_pages')->onDelete('CASCADE');
         $t->foreign('snapshot_by')->references('id')->on('users')->onDelete('CASCADE');
